% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-clean.R
\name{data_clean}
\alias{data_clean}
\title{Basic cleaning of data}
\usage{
data_clean(data.xls)
}
\arguments{
\item{data.xls}{data frame output from \code{\link{xls_read}}}
}
\value{
data frame with the same values as the input data, but colnames are
normalized and dates are parsed.
}
\description{
Clean variable names and parse date
}
\details{
This function parse dates, nomalize 'variable names' appending
the hour after each one. Date column is added in last column.
}
\examples{
\dontrun{
if(interactive()){
#xfiles_l <- list.files("vignettes/dvd_xls_files", recursive = TRUE, full.names = TRUE)
#xls_file <- xfiles_l[1]
#raw_data <- xls_read(xls_file)
#clean_data <- data_clean(raw_data)
#str(clean_data)
 }
}
}
