% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_xls_files.R
\name{find_xls_files}
\alias{find_xls_files}
\title{Find the pair of Excel files from a automatic weather station}
\usage{
find_xls_files(file.name, verbose = TRUE)
}
\arguments{
\item{file.name}{character vector with paths to Excel files;
in general a list with path to file 2 (for more details see the vignette:
\code{vignette("data-org", package = "rinmetxls")})}

\item{verbose}{logical, should print messages? Default: TRUE}
}
\value{
character vector of file paths; it is expected two file paths
}
\description{
Find the pair of Excel files from a automatic weather station
}
\examples{
\dontrun{
if(interactive()){
  p <- system.file("extdata/dvd_xls_files", package = "rinmetxls")
  pf <- list.files(
    p,
    pattern = ".*[[:punct:]]_\\\\.xls",
    recursive = TRUE,
    full.names = TRUE
  )
 find_xls_files(pf[1])
 find_xls_files(pf[2])
 find_xls_files(pf[3])
 }
}
}
\concept{file functions}
