% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_ffilename.R
\name{metadata_ffilename}
\alias{metadata_ffilename}
\title{Parse AWS metadata from Excel file name}
\usage{
metadata_ffilename(file.name)
}
\arguments{
\item{file.name}{A character vector of one or more file paths.}
}
\value{
a data frame with columns:
\itemize{
\item uf, Federative Unit
\item name, station name
\item id, station identification code
}
}
\description{
Parse Station name, identification code and UF from Excel file
name
}
\details{
\code{metadata_ffilename()} get metadata from file name. It is
usefull to compare with metadata obtained from the header in Excel file
}
\examples{
\dontrun{
if(interactive()){
 fname <- "dvd_xls_files/NORTE_2017/AM_2017/__AM_A134_S._G._DA_CACHOEIRA_.xls.xls"
 metadata_ffilename(fname)
 }
}
}
