% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xls-read.R
\name{metadata_join}
\alias{metadata_join}
\title{Join metadata obtained from different sources}
\usage{
metadata_join(path.file, data.xls)
}
\arguments{
\item{path.file}{path to Excel file}

\item{data.xls}{data frame}
}
\value{
data frame with columns:
\itemize{
\item name, station name obtained from header in file
\item uf, federative unit obtained from header in file
\item lon, longitude (decimal degrees)
\item lat, latitude (decimal degrees)
\item uf_ffname, federative unit obtained from file name
\item name_ffname, station name obtained from file name
\item id, station identification code obtained from file name
}
}
\description{
Join metadata obtained from (i) the name of the Excel file and
(ii) the header in the Excel file.
}
\seealso{
Other metadata functions: \code{\link{metadata_ffilename}},
  \code{\link{metadata_parse}}
}
\concept{metadata functions}
