% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-parse.R
\name{metadata_parse}
\alias{metadata_parse}
\title{Parse metadata from header in the Excel file}
\usage{
metadata_parse(data.xls)
}
\arguments{
\item{data.xls}{data frame imported by \code{\link{xls_read}}}
}
\value{
data frame with columns:
\itemize{
\item name: station name
\item uf: federative unit
\item lon: longitude in decimal degrees
\item lat: latitude in decimal degrees
\item alt: altitude in m
}
}
\description{
Parse metadata from header in the Excel file
}
\details{
This function is used in \code{\link{xls_read}} to extract metadata
from the header in the Excel file.
}
\seealso{
Other metadata functions: \code{\link{metadata_ffilename}},
  \code{\link{metadata_join}}
}
\concept{metadata functions}
