% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xls-read.R
\name{ncols_by_variable}
\alias{ncols_by_variable}
\title{Number of columns for each variable in data}
\usage{
ncols_by_variable(data.xls)
}
\arguments{
\item{data.xls}{is the output data frame from \code{\link{xls_read}}}
}
\value{
data frame
}
\description{
Count the number of columns for each variable found in Excel
data file
}
\details{
This function is used to check whether the number of columns in
each variable in the Excel file are the same for the same data type (1 or 2).
It used as input the output of the function \code{\link{xls_read}}.
This identifies non-standard files of data arrangement.
}
\examples{
\dontrun{
if(interactive()){
library(stringr)
xfiles_l <- list.files("vignettes/dvd_xls_files", recursive = TRUE, full.names = TRUE)
fpath <- basename(xfiles_l)
length(xfiles_l)
# select file 2
sel <- str_detect(fpath, ".*_(\\\\S|\\\\s)\\\\.xls\\\\.xls") |
  str_detect(fpath, ".*[:punct:]_\\\\.xls") |
  str_detect(fpath, ".*[A-z]{1}_\\\\.xls\\\\.xls") |
  str_detect(fpath, ".*[A-z]{1}_V\\\\.xls\\\\.xls")
length(xfiles_l[sel])
# verficação dos arquivos 2 (outras variáveis)
data_l <- lapply(xfiles_l[sel], function(ifile) xls_read(file.xls = ifile))
# check num od columns by variable
ncol_vars <- plyr::ldply(l, function(x) {
  cat(attr(x, "meta")[["id"]], "\\n")
  ncols_by_variable(x)
})
head(ncol_vars)
 }
}
}
\concept{data-check functions}
