% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_coords.R
\name{parse_coords}
\alias{parse_coords}
\title{Parse lon, lat, alt from data read from a Excel file}
\usage{
parse_coords(xlsdf)
}
\arguments{
\item{xlsdf}{data frame imported by \code{read_aws_xls_file()}}
}
\value{
data frame with AWS coordinates`
}
\description{
Parse lon, lat, alt from data read from a Excel file
}
