% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize_varnames.R
\name{sanitize_varname}
\alias{sanitize_varname}
\title{Sanitize variable names by removing invalid and reserved characters}
\usage{
sanitize_varname(vnames, sep = "_", no_accent = TRUE)
}
\arguments{
\item{vnames}{A character vector to be sanitized}

\item{sep}{A character vector used to replace invalid characters,
Default: '_'}

\item{no_accent}{logical, if is to remove accents from vnames, Default: TRUE}
}
\value{
A character vector in lowercase and a underscore (sep) to separate
nouns
}
\description{
Make a reasonable attempt of converting variables names from a
dataset into a preferred standard form
}
\details{
\code{sanitize_varname()} removes or replace the following:
\itemize{
\item \href{https://en.wikipedia.org/wiki/C0_and_C1_control_codes}{Control characters}
\item \href{https://kb.acronis.com/content/39790}{Reserved characters}
\item Unix reserved filenames (\code{.} and \code{..})
\item Trailing periods and spaces (invalid on Windows)
\item Windows reserved filenames (\code{CON}, \code{PRN}, \code{AUX}, \code{NUL}, \code{COM1}, \code{COM2},
\code{COM3}, COM4, \code{COM5}, \code{COM6}, \code{COM7}, \code{COM8}, \code{COM9}, \code{LPT1}, \code{LPT2},
\code{LPT3}, \code{LPT4}, \code{LPT5}, \code{LPT6}, LPT7, \code{LPT8}, and \code{LPT9})
\item any accented noum or punctuation
\item any resulting initial or trailing underscore or multiples
\item uppercase to lowercase
\item repeated seperator
\item names starting with a number by a character
}
}
\examples{
sanitize_varname(c("esúpido", "^ ãb ", "..c`a§", "A .xls.xls"))
}
