% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_distinct.R
\name{str_distinct}
\alias{str_distinct}
\title{Extract characters that differ between two strings}
\usage{
str_distinct(a, b)
}
\arguments{
\item{a}{character vector}

\item{b}{character vector}
}
\value{
character vector with strings in a that are not in b
}
\description{
Extract characters that differ between two strings
}
\examples{
\dontrun{
if(interactive()){
 str_distinct("abc", "cde")
 str_distinct("AERRRTX", "TRRA")
 }
}
}
\seealso{
\code{\link[data.table]{data.table-package}}
}
