% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{tidy_data}
\alias{tidy_data}
\title{Tidy data from Excel file}
\usage{
tidy_data(data.clean)
}
\arguments{
\item{data.clean}{data frame processed by \code{\link{data_clean}}}
}
\value{
data frame in tidy format
}
\description{
Tidy hourly data from AWS Excel file
}
\details{
This function restructures the data in the tidy format,
standardizes the name of the variables, adds the AWS identifier code
to a new variable named "site".
The tidy format is such that every value belongs to a variable and
an observation. Please refer to
\href{https://cran.r-project.org/web/packages/tidyr/vignettes/tidy-data.html}{tidy data}
for more details.
}
\examples{
\dontrun{
if(interactive()){
xfiles_l <- list.files(
  "vignettes/dvd_xls_files",
  recursive = TRUE,
  full.names = TRUE
)
xls_file <- sample(xfiles_l, 1)
data_tidy <- tidy_data(data_clean(xls_read(xls_file)))
str(data_tidy)
#View(data_tidy)
 }
}
}
\seealso{
\code{\link{xls_read}}, \code{\link{data_clean}}

Other data processing: \code{\link{data_clean}},
  \code{\link{varnames_recode}}
}
\concept{data processing}
