% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tidy.R
\name{varnames_possible}
\alias{varnames_possible}
\title{Collect all possible names in Excel data files}
\usage{
varnames_possible(paths.xls, .verbose = TRUE)
}
\arguments{
\item{paths.xls}{character vector, file paths to Excel files}

\item{.verbose}{logical, should print messages? Default: TRUE}
}
\value{
character vector with the unique variable names found in Excel files
}
\description{
Collect all possible names in Excel data files
}
\examples{
\dontrun{
if(interactive()){
xfiles_l <- list.files(
  "vignettes/dvd_xls_files",
  recursive = TRUE,
  full.names = TRUE
)
xls_paths <- sample(xfiles_l, 2)
xls_vnames <- varnames_possible(xls_paths)
xls_vnames
 }
}

}
\concept{data check}
