% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{varnames_recode}
\alias{varnames_recode}
\title{Recode variable names}
\usage{
varnames_recode(varnames)
}
\arguments{
\item{varnames}{variable names from data}
}
\description{
Recode variable names
}
\seealso{
Other data processing: \code{\link{data_clean}},
  \code{\link{tidy_data}}
}
\concept{data processing}
