% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xls-import.R
\name{xls_import}
\alias{xls_import}
\title{Import, clean and tidy data from a Excel file}
\usage{
xls_import(file.xls, verbose = TRUE, na.strings = "NULL",
  kjm2.wm2 = (1000/10^6)/0.0864 * 24)
}
\arguments{
\item{file.xls}{Import data from a Excel file}

\item{verbose}{
    a logical indicating if some \dQuote{progress report} should be given.
  }

\item{na.strings}{Character vector of strings to interpret as missing
values. By default, readxl treats blank cells as missing data.
Default: 'NULL'.}

\item{kjm2.wm2}{unit conversion for solar incoming radiation,
Default: (1000/10^6)/0.0864 * 24. Set equal to NULL to not apply unit
conversion to W m^-2^.}
}
\value{
data frame with tidy data
}
\description{
Import, clean and tidy data from a Excel file
}
\details{
This function import and tidy data from a Excel data file.
}
\examples{
\dontrun{
if(interactive()){
xfiles_l <- list.files(
  "vignettes/dvd_xls_files",
  recursive = TRUE,
  full.names = TRUE
)
xls_file <- sample(xfiles_l, 1)
aws_data <- xls_import(file.xls = xls_file)
str(aws_data)
attr(aws_data, "meta")
 }
}
}
\seealso{
\code{\link{xls_read}}, \code{\link{data_clean}}, \code{\link{data_tidy}}
}
