% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xls-read.R
\name{xls_read}
\alias{xls_read}
\title{Read and select data}
\usage{
xls_read(file.xls, na = "NULL", verbose = TRUE)
}
\arguments{
\item{file.xls}{character, path to Excel file}

\item{na}{Character vector of strings to interpret as missing values. By
default, readxl treats blank cells as missing data.}

\item{verbose}{
    a logical indicating if some \dQuote{progress report} should be given.
  }
}
\value{
data frame
}
\description{
This function Reads Excel data file, determine boundaries of
the table (body data), get metadata from AWS and add it as a attribute of
output data frame
}
\details{
xls_read() read the data in the excel file and determine
the limits of the data table, that is, the body of the data (excludes header
with AWS metadata).

The initial row of data (line with name of the variables) is detected by
searching for the line with at least 3 letters, in order to find the name of
the meteorological variables. This procedure was adopted because depending
on the data request to INMET the file may contain variables in order and with
different names.

Automatic meteorological station (AWS) metadata will be extracted from two
sources: (i) the name of the Excel file and (ii) the header in the Excel
file. The metadata is included as an attribute named 'meta' in the data frame with the
data imported from the Excel file. You can access metada with
\code{attr(raw_data, 'meta')}.
}
\examples{
\dontrun{
xls_list <- list.files(
  system.file("vignettes/dvd_xls_files", package = "rinmetxls"),
  full.names = TRUE, recursive = TRUE
)
xls_file <- grep("BELEM", xls_list, value = TRUE)[1]
if (file.exists(xls_file)) {
  raw_data <- xls_read(xls_file)
  str(raw_data)
}
}
}
\seealso{
\code{\link[readxl]{read_excel}}
}
