% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-parse.R
\name{metadata_parse}
\alias{metadata_parse}
\title{Parse metadata from header in the Excel file}
\usage{
metadata_parse(xlsdf)
}
\arguments{
\item{xlsdf}{data frame imported by \code{read_aws_xls_file()}}
}
\value{
data frame with columns:
\itemize{
\item name: station name
\item uf: federative unit
\item lon: longitude in decimal degrees
\item lat: latitude in decimal degrees
\item alt: altitude in m
}
}
\description{
Parse metadata from header in the Excel file
}
