% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize_varnames.R
\name{sanitize_varname}
\alias{sanitize_varname}
\title{Sanitize variable names by removing invalid and reserved characters}
\usage{
sanitize_varname(vnames, sep = "_", no_accent = TRUE)
}
\arguments{
\item{vnames}{A character vector to be sanitized}

\item{sep}{A character vector used to replace invalid characters,
Default: '_'}

\item{no_accent}{logical, if is to remove accents from vnames, Default: TRUE}
}
\value{
A character vector in lowercase and a underscore (sep) to separate
nouns
}
\description{
Make a reasonable attempt of converting variables names from a
dataset into a preferred standard form
}
\details{
`sanitize_varname()` removes or replace the following:

- [Control characters](https://en.wikipedia.org/wiki/C0_and_C1_control_codes)

- [Reserved characters](https://kb.acronis.com/content/39790)

- Unix reserved filenames (`.` and `..`)

- Trailing periods and spaces (invalid on Windows)

- Windows reserved filenames (`CON`, `PRN`, `AUX`, `NUL`, `COM1`, `COM2`,
  `COM3`, COM4, `COM5`, `COM6`, `COM7`, `COM8`, `COM9`, `LPT1`, `LPT2`,
  `LPT3`, `LPT4`, `LPT5`, `LPT6`, LPT7, `LPT8`, and `LPT9`)

- any accented noum or punctuation

- any resulting initial or trailing underscore or multiples

- uppercase to lowercase

- repeated seperator

- names starting with a number by a character
}
\examples{
sanitize_varname(c("esúpido", "^ ãb ", "..c`a§", "A .xls.xls"))
}
